+++ resources/js/Components/VideoPlayer/Chat/ChatMessage.vue
@@ -7,7 +7,7 @@
                 src="" class="rounded-full h-8 w-8 object-cover bg-gray-300">
        </div>
         <div class="flex flex-col">
-            <div :class="messageBgClass"
+            <div :class="[pipMessageBgClass, messageBgClass]"
                  class="flex flex-col rounded-l-xl rounded-r-xl px-2 pb-1 bg-opacity-50 break-words">
                 <span class="text-xs font-semibold text-gray-100 pt-1">{{ message.user_name }}</span>
                 <span class="text-white break-words" v-html="message.message" />
@@ -22,10 +22,12 @@
 <script setup>
 import { computed } from "vue";
 import { useAppSettingStore } from '@/Stores/AppSettingStore';
+import { useVideoPlayerStore } from '@/Stores/VideoPlayerStore';
 import { useChatStore } from "@/Stores/ChatStore";
 import { useTimeAgo } from '@vueuse/core';
 
 const appSettingStore = useAppSettingStore();
+const videoPlayerStore = useVideoPlayerStore();
 let chatStore = useChatStore()
 
 let props = defineProps({
@@ -35,6 +37,10 @@ let props = defineProps({
 const timeAgo = useTimeAgo(props.message.created_at)
 
 const messageBgClass = computed(() => {
+    return videoPlayerStore.fullPage ? 'lg:bg-gray-800 lg:bg-opacity-60' : '';
+});
+
+const pipMessageBgClass = computed(() => {
     return appSettingStore.pipChatMode
         ? appSettingStore.chatMessageBgColor
         : appSettingStore.primaryChatMessageBgColor;