+++ resources/js/Components/VideoPlayer/OttFullPageDisplay/Channels.vue
@@ -9,7 +9,7 @@
     >
 
 
-            <div v-if="videoPlayerStore.ottChannels" class="channelsFullPageContainer">
+            <div v-if="videoPlayerStore.ottChannels" class="channelsFullPageContainer hide-scrollbar">
                 <div class="channelsFullPageInner">
                     <div class="flex flex-col p-5 mt-2">
                         <div class="text-3xl text-center font-semibold uppercase mb-3 w-full bg-green-900 text-white p-2">CHANNELS</div>