+++ resources/js/Components/VideoPlayer/OttFullPageDisplay/Playlist.vue
@@ -11,7 +11,7 @@
 
 <!--        <div v-if="videoPlayerStore.showPlaylist" class="playlistForStreamPageStandard w-96 max-w-[12rem] bottom-0 bg-opacity-80 bg-orange-600 text-sm text-white pb-12-->
 <!--                chat-mask overflow-y-auto scroll-smooth hover:scroll-auto break-words">-->
-        <div v-if="videoPlayerStore.ottPlaylist" class="playlistFullPageContainer">
+        <div v-if="videoPlayerStore.ottPlaylist" class="playlistFullPageContainer hide-scrollbar">
 
 <!--        Previous version (full screen background for Ott) -->
 <!--        <div v-if="videoPlayerStore.showChannels" class="channelsForStreamPageStandard max-w-screen h-screen left-0 bottom-0 sm:bottom-8 bg-opacity-80 bg-green-600 text-sm text-white pb-12-->